refactor: restructure as blue-onnx package for PyPI publishing
Move _common.py and _blue_vocab.py into src/blue_onnx/ so the package
is self-contained. Rename project from 'blue' to 'blue-onnx' to match
the intended PyPI name. blue_pt and blue_trt now import shared code
from blue_onnx instead of the src root.
